refactor: improve exception handling and validation logic

+++ backend/src/main/java/com/smarthire/resume/service/CandidatoService.java
@@ -41,21 +41,14 @@ public class CandidatoService {
 
     @Transactional
     public Candidato salvar(Candidato candidato) {
-        boolean emailEmUso = candidatoRepository.findByEmail(candidato.getEmail())
-                .filter(e-> !e.equals(candidato))
-                .isPresent();
-        if (emailEmUso) {
-            throw new BusinessRuleException("E-mail já cadastrado no sistema.");
-        }
+        verificarEmailEmUso(candidato.getEmail(), candidato);
         return candidatoRepository.save(candidato);
     }
 
     public CandidatoDto listarCandidatos(Candidato candidato) {
         Vaga vaga = candidato.getVaga();
         Curriculo curriculo = candidato.getCurriculo();
-        if(curriculo == null) {
-            throw new BusinessRuleException("Candidato não possui currículo");
-        }
+        verificarNulidadeCurriculo(curriculo);
 
         VagaResumoDto vagaResumoDto = null;
         if (vaga != null) {
@@ -111,7 +104,8 @@ public class CandidatoService {
                 .orElseThrow(() -> new BusinessRuleException("Vaga não encontrada"));
 
         
-
+        verificarNulidadeCurriculo(curriculo);
+        verificarVagaAtiva(vaga);
         candidato.atualizarCom(data, curriculo, vaga);
 
         return candidatoRepository.save(candidato);
@@ -124,6 +118,9 @@ public class CandidatoService {
         }
         Candidato candidato = candidatoRepository.findById(id)
                 .orElseThrow(() -> new BusinessRuleException("Candidato não encontrado."));
+        if(candidato.getVaga() != null) {
+            throw new BusinessRuleException("Candidato não pode ser deletado, pois está vinculado a uma vaga.");
+        }
         candidatoRepository.delete(candidato);
     }
 
@@ -131,6 +128,7 @@ public class CandidatoService {
     @Transactional
     public void criarComCurriculo(Curriculo curriculo, UUID idVaga) {
         Candidato candidato = new Candidato();
+        verificarNulidadeCurriculo(curriculo);
         candidato.setCurriculo(curriculo);
         candidato.setNome(curriculo.getNome());
         candidato.setEmail(curriculo.getEmail());
@@ -151,10 +149,27 @@ public class CandidatoService {
         if(candidato.getVaga() != null) {
             throw new BusinessRuleException("Candidato já está vinculado a uma vaga.");
         }
-        if(!vaga.isActive()) {
-            throw new BusinessRuleException("Essa vaga não está aberta a candidaturas");
-        }
+        verificarVagaAtiva(vaga);
         candidato.setVaga(vaga);
         candidatoRepository.save(candidato);
+    }
+
+    private void verificarNulidadeCurriculo(Curriculo curriculo) {
+        if(curriculo == null) {
+            throw new BusinessRuleException("Currículo não pode ser nulo.");
+        }
+    }
+    private void verificarEmailEmUso(String email, Candidato candidato) {
+        boolean emailEmUso = candidatoRepository.findByEmail(candidato.getEmail())
+                .filter(e-> !e.equals(candidato))
+                .isPresent();
+        if (emailEmUso) {
+            throw new BusinessRuleException("E-mail já cadastrado no sistema.");
+        }
+    }
+    private void verificarVagaAtiva(Vaga vaga) {
+        if(!vaga.isActive()) {
+            throw new BusinessRuleException("Essa vaga não está aberta a candidaturas.");
+        }
     }
 }
